third_party/WebKit/LayoutTests/editing/deleting/delete-leading-ws-001.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../assert_selection.js"></script>
<script>
    test(() => assert_selection(
        '<div contenteditable>f   ^baz|</div>',
        'delete',
        '<div contenteditable>f\u{00A0}|</div>'),
        'Delete then collapse whitespaces to nbsp');
</script>
